refactor(hooks): rename cwd_tracking to cwd_awareness
Following the naming convention of time_awareness, token_awareness,
cost_awareness etc. The module injects awareness hints about CWD changes
rather than tracking changes itself — "awareness" better reflects its role.

- Rename cwd_tracking.py → cwd_awareness.py
- Rename test_cwd_tracking.py → test_cwd_awareness.py
- Update hook registration name: cwd_tracking.notification → cwd_awareness.notification
- Update all imports and references
